1.0

+++ 0x01-variables_if_else_while/0-positive_or_negative.c
@@ -11,13 +11,13 @@ int main(void)
 
 	scanf("%lld", &number);
 	if (number > 0)
-		printf("%lld: is positive\n", number);
+		printf("%lld: is positive\n", (long long)number);
 
 	else if (number == 0)
-		printf("%lld: is zero\n", number);
+		printf("%lld: is zero\n", (long long)number);
 
 	else
-		printf("%lld: is negative\n", number);
+		printf("%lld: is negative\n", (long long)number);
 
 	return (0);
 }
